Decompose extensions API object getters to a separate source file.
